Edits to README.md
Edited the "highly customizable and flexible..." bullet point of features section. Changed "...a node editor and custom scripting language, named Maxim" to "...a node editor and Maxim, a custom scripting language." The edit clarifies that it's only the custom scripting language that is called Maxim (if that is what you meant).

Edited presentation of instructions in building backend. The two separate instructions are in bullets and are each preceded by a bolded title . It is now clear to the reader that there are two options in building backends.